StateManager: extract generic cache functionality to superclass Cache
