+++ packages/statemanager/src/cache/index.ts
@@ -1,1 +1,2 @@
 export * from './account'
+export * from './types'
