+++ packages/statemanager/src/cache/types.ts
@@ -1,0 +1,9 @@
+export enum CacheType {
+  LRU = 'lru',
+  ORDERED_MAP = 'ordered_map',
+}
+
+export interface CacheOpts {
+  size: number
+  type: CacheType
+}
